requirements.txt
dj-database-url==0.5.0
Django==2.1.5
django-money==0.14.3
django-rest-framework==0.1.0
psycopg2==2.7.6.1
python-dotenv==0.10.1
pytz==2018.7
